+++ stone-js-cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.8.13
+
+### Patch Changes
+
+- @stone-js/core@0.8.13
+- @stone-js/pipeline@0.8.13
+- @stone-js/config@0.8.13
+- @stone-js/filesystem@0.8.13
+- @stone-js/router@0.8.13
+- @stone-js/node-cli-adapter@0.8.13
+
 ## 0.8.12
 
 ### Patch Changes