+++ stone-js-edge-adapter/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stone-js/edge-adapter
 
+## 0.8.13
+
+### Patch Changes
+
+- @stone-js/fetch-adapter@0.8.13
+- @stone-js/core@0.8.13
+
 ## 0.8.12
 
 ### Patch Changes