+++ stone-js-use-react-core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stone-js/use-react-core
 
+## 0.8.13
+
+### Patch Changes
+
+- Updated dependencies [8c2b600]
+  - @stone-js/http-core@0.8.13
+  - @stone-js/core@0.8.13
+  - @stone-js/config@0.8.13
+  - @stone-js/router@0.8.13
+  - @stone-js/browser-core@0.8.13
+  - @stone-js/use-view@0.8.13
+
 ## 0.8.12
 
 ### Patch Changes